Gyrot fungerar isch

+++ Sensor/main/read.h
@@ -3,6 +3,6 @@
 
 uint8_t read_reflex();
 uint8_t read_IR();
-uint8_t read_gyro();
+int read_gyro();
 
 #endif
